chore: cursor long타입으로 변경

+++ src/main/java/org/sopt/makers/internal/repository/WordChainGameQueryRepository.java
@@ -17,7 +17,7 @@ public class WordChainGameQueryRepository {
     private final JPAQueryFactory queryFactory;
 
     public List<WordChainGameRoom> findAllLimitedGameRoom(
-            Integer limit, Integer cursor
+            Integer limit, Long cursor
     ) {
         val room = QWordChainGameRoom.wordChainGameRoom;
         return queryFactory.selectFrom(room)
